Added a virtual Environment;  Corrected SQL install. Updated python. Updated pip. Installed homebrew. Installed Django. Added C++, Corrected link $PATH, effectively installed pygame, and requirements.txt finally
I was able to appropriately set up the virtual enviornment in the right place. I had difficulty translating the language to work with macos and visualcode (as the original creator has a windows and used cygwin.
There was also an issue getting the originally proposed set-up because mysqlclient and macos and home brew are currently not wanted to work together correctly. Luckidly I went back to the originary repo and found the issues notes, and found some notes for macos functionality. (thank goodness).
 https://github.com/blreams/collegefootballpick10/issues/1. I was also able to correctly use $PATH, and other terminal commands to move and create some files.
I had been having issues with the terminal reading under an older version of python, which is another reason the code wasn't running) -- - but after hours of reserach and and great professor: I was able to commit to my master branch.
Here is the link to the holy grail for this project. https://github.com/blreams/collegefootballpick10/search?q=mysql&type=Issues
 > I was finally able to correct execute the required installs.

Co-Authored-By: msyed23 <msyed23@users.noreply.github.com>